+++ src/widgets/draggable_tab_bar/node.rs
@@ -67,6 +67,9 @@ pub struct DraggableTabBarNode {
     pub disabled: bool,
     pub disabled_style: Style,
     pub focusable: bool,
+    pub tab_stop: bool,
+    pub on_focus: Option<Callback<()>>,
+    pub on_blur: Option<Callback<()>>,
 }
 
 impl DraggableTabBarNode {
@@ -98,6 +101,18 @@ impl DraggableTabBarNode {
 impl WidgetNode for DraggableTabBarNode {
     fn is_focusable(&self) -> bool {
         self.focusable
+    }
+
+    fn is_tab_stop(&self) -> bool {
+        self.focusable && self.tab_stop
+    }
+
+    fn on_focus_callback(&self) -> Option<&Callback<()>> {
+        self.on_focus.as_ref()
+    }
+
+    fn on_blur_callback(&self) -> Option<&Callback<()>> {
+        self.on_blur.as_ref()
     }
 
     fn has_on_click(&self) -> bool {
@@ -180,6 +195,9 @@ impl From<DraggableTabBar> for DraggableTabBarNode {
             disabled: value.disabled,
             disabled_style: value.disabled_style,
             focusable: value.focusable,
+            tab_stop: value.tab_stop,
+            on_focus: value.on_focus,
+            on_blur: value.on_blur,
         }
     }
 }